feat: filter moved as comp
